@@ -1,16 +1,19 @@
 import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+
 import {
   Text,
   View,
   Image,
   Animated,
-  ScrollView,
-  StyleSheet
+  ScrollView
 } from 'react-native';
 
 import { Icon, List, ListItem } from 'react-native-elements';
 
 import { USER, FACEBOOK_LIST, SLACK_LIST, GENERIC_LIST, SCREEN_WIDTH, SCREEN_HEIGHT, DEFAULT_WINDOW_MULTIPLIER, DEFAULT_NAVBAR_HEIGHT } from './constants';
+
+import styles from './styles';
 
 const ScrollViewPropTypes = ScrollView.propTypes;
 
@@ -23,8 +26,13 @@ export default class ParallaxScrollView extends Component {
     };
   }
 
+  scrollTo(where) {
+    if (!this._scrollView) return;
+    this._scrollView.scrollTo(where);
+  }
+
   renderBackground() {
-    var { windowHeight, backgroundSource } = this.props;
+    var { windowHeight, backgroundSource, onBackgroundLoadEnd, onBackgroundLoadError } = this.props;
     var { scrollY } = this.state;
     if (!windowHeight || !backgroundSource) {
       return null;
@@ -53,25 +61,28 @@ export default class ParallaxScrollView extends Component {
           }
         ]}
         source={backgroundSource}
+        onLoadEnd={onBackgroundLoadEnd}
+        onError={onBackgroundLoadError}
       >
       </Animated.Image>
     );
   }
 
   renderHeaderView() {
-    var { windowHeight, backgroundSource, userImage, userName, userTitle } = this.props;
-    var { scrollY } = this.state;
+    const { windowHeight, backgroundSource, userImage, userName, userTitle, navBarHeight } = this.props;
+    const { scrollY } = this.state;
     if (!windowHeight || !backgroundSource) {
       return null;
     }
 
-    const newWindowHeight = windowHeight - DEFAULT_NAVBAR_HEIGHT;
+    const newNavBarHeight = navBarHeight || DEFAULT_NAVBAR_HEIGHT;
+    const newWindowHeight = windowHeight - newNavBarHeight;
 
     return (
       <Animated.View
         style={{
           opacity: scrollY.interpolate({
-            inputRange: [-windowHeight, 0, windowHeight * DEFAULT_WINDOW_MULTIPLIER + DEFAULT_NAVBAR_HEIGHT],
+            inputRange: [-windowHeight, 0, windowHeight * DEFAULT_WINDOW_MULTIPLIER + newNavBarHeight],
             outputRange: [1, 1, 0]
           })
         }}
@@ -98,8 +109,8 @@ export default class ParallaxScrollView extends Component {
   }
 
   renderNavBarTitle() {
-    var { windowHeight, backgroundSource, navBarTitleColor } = this.props;
-    var { scrollY } = this.state;
+    const { windowHeight, backgroundSource, navBarTitleColor, navBarTitleComponent } = this.props;
+    const { scrollY } = this.state;
     if (!windowHeight || !backgroundSource) {
       return null;
     }
@@ -113,86 +124,108 @@ export default class ParallaxScrollView extends Component {
           })
         }}
       >
+        {navBarTitleComponent ||
         <Text style={{ fontSize: 18, fontWeight: '600', color: navBarTitleColor || 'white' }}>
           {this.props.navBarTitle || USER.name}
-        </Text>
+        </Text>}
       </Animated.View>
     );
   }
 
   rendernavBar() {
-    var {
+    const {
       windowHeight, backgroundSource, leftIcon,
-      rightIcon, leftIconOnPress, rightIconOnPress, navBarColor
+      rightIcon, leftIconOnPress, rightIconOnPress, navBarColor, navBarHeight, leftIconUnderlayColor, rightIconUnderlayColor
     } = this.props;
-    var { scrollY } = this.state;
+    const { scrollY } = this.state;
     if (!windowHeight || !backgroundSource) {
       return null;
     }
 
-    return (
-      <Animated.View
-        style={{
-          height: DEFAULT_NAVBAR_HEIGHT,
-          width: SCREEN_WIDTH,
-          flexDirection: 'row',
-          backgroundColor: scrollY.interpolate({
-            inputRange: [-windowHeight, windowHeight * DEFAULT_WINDOW_MULTIPLIER, windowHeight * 0.8],
-            outputRange: ['transparent', 'transparent', navBarColor || 'rgba(0, 0, 0, 1.0)']
-          })
-        }}
-      >
-      {leftIcon &&
-        <View
-          style={{
-            flex: 1,
-            marginTop: 20,
-            justifyContent: 'center',
-            alignItems: 'center'
-          }}
-        >
-          <Icon
-            name={leftIcon && leftIcon.name || 'menu'}
-            type={leftIcon && leftIcon.type || 'simple-line-icon'}
-            color={leftIcon && leftIcon.color || 'white'}
-            size={leftIcon && leftIcon.size || 23}
-            onPress={leftIconOnPress}
-            underlayColor='transparent'
-          />
-        </View>
-      }
-        <View
-          style={{
-            flex: 5,
-            marginTop: 20,
-            justifyContent: 'center',
-            alignItems: 'center',
-            alignSelf: 'center'
-          }}
-        >
-          {this.renderNavBarTitle()}
-        </View>
-      {rightIcon &&
-        <View
-          style={{
-            flex: 1,
-            marginTop: 20,
-            justifyContent: 'center',
-            alignItems: 'center'
-          }}
-        >
-          <Icon
-            name={rightIcon && rightIcon.name || 'present'}
-            type={rightIcon && rightIcon.type || 'simple-line-icon'}
-            color={rightIcon && rightIcon.color || 'white'}
-            size={rightIcon && rightIcon.size || 23}
-            onPress={rightIconOnPress}
-            underlayColor='transparent'
-          />
-        </View>
-      }
-      </Animated.View>
-    );
+    const newNavBarHeight = navBarHeight || DEFAULT_NAVBAR_HEIGHT;
+
+    if(this.props.navBarView)
+    {
+        return (
+          <Animated.View
+            style={{
+              height: newNavBarHeight,
+              width: SCREEN_WIDTH,
+              flexDirection: 'row',
+              backgroundColor: scrollY.interpolate({
+                inputRange: [-windowHeight, windowHeight * DEFAULT_WINDOW_MULTIPLIER, windowHeight * 0.8],
+                outputRange: ['transparent', 'transparent', navBarColor || 'rgba(0, 0, 0, 1.0)'],
+                extrapolate: 'clamp'
+              })
+            }}
+          >
+          {this.props.navBarView}
+          </Animated.View>
+        );
+    }
+    else
+    {
+        return (
+          <Animated.View
+            style={{
+              height: newNavBarHeight,
+              width: SCREEN_WIDTH,
+              flexDirection: 'row',
+              backgroundColor: scrollY.interpolate({
+                inputRange: [-windowHeight, windowHeight * DEFAULT_WINDOW_MULTIPLIER, windowHeight * 0.8],
+                outputRange: ['transparent', 'transparent', navBarColor || 'rgba(0, 0, 0, 1.0)']
+              })
+            }}
+          >
+          {leftIcon &&
+            <View
+              style={{
+                flex: 1,
+                justifyContent: 'center',
+                alignItems: 'center'
+              }}
+            >
+              <Icon
+                name={leftIcon && leftIcon.name || 'menu'}
+                type={leftIcon && leftIcon.type || 'simple-line-icon'}
+                color={leftIcon && leftIcon.color || 'white'}
+                size={leftIcon && leftIcon.size || 23}
+                onPress={leftIconOnPress}
+                underlayColor={leftIconUnderlayColor || 'transparent'}
+              />
+            </View>
+          }
+            <View
+              style={{
+                flex: 5,
+                justifyContent: 'center',
+                alignItems: 'center',
+                alignSelf: 'center'
+              }}
+            >
+              {this.renderNavBarTitle()}
+            </View>
+          {rightIcon &&
+            <View
+              style={{
+                flex: 1,
+                justifyContent: 'center',
+                alignItems: 'center'
+              }}
+            >
+              <Icon
+                name={rightIcon && rightIcon.name || 'present'}
+                type={rightIcon && rightIcon.type || 'simple-line-icon'}
+                color={rightIcon && rightIcon.color || 'white'}
+                size={rightIcon && rightIcon.size || 23}
+                onPress={rightIconOnPress}
+                underlayColor={rightIconUnderlayColor || 'transparent'}
+              />
+            </View>
+          }
+          </Animated.View>
+        );
+    }
   }
 
   renderTodoListContent() {
@@ -245,7 +278,7 @@ export default class ParallaxScrollView extends Component {
   }
 
   render() {
-    var { style, ...props } = this.props;
+    const { style, ...props } = this.props;
 
     return (
       <View style={[styles.container, style]}>
@@ -281,55 +314,16 @@ ParallaxScrollView.defaultProps = {
 
 ParallaxScrollView.propTypes = {
   ...ScrollViewPropTypes,
-  backgroundSource: Image.propTypes.source,
-  windowHeight: React.PropTypes.number,
-  navBarTitle: React.PropTypes.string,
-  navBarTitleColor: React.PropTypes.string,
-  navBarColor: React.PropTypes.string,
-  userImage: React.PropTypes.string,
-  userName: React.PropTypes.string,
-  userTitle: React.PropTypes.string,
-  headerView: React.PropTypes.node,
-  leftIcon: React.PropTypes.object,
-  rightIcon: React.PropTypes.object
+  backgroundSource: PropTypes.object,
+  windowHeight: PropTypes.number,
+  navBarTitle: PropTypes.string,
+  navBarTitleColor: PropTypes.string,
+  navBarTitleComponent: PropTypes.node,
+  navBarColor: PropTypes.string,
+  userImage: PropTypes.string,
+  userName: PropTypes.string,
+  userTitle: PropTypes.string,
+  headerView: PropTypes.node,
+  leftIcon: PropTypes.object,
+  rightIcon: PropTypes.object
 };
-
-var styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    borderColor: 'transparent'
-  },
-  scrollView: {
-    backgroundColor: 'transparent'
-  },
-  background: {
-    position: 'absolute',
-    backgroundColor: '#2e2f31',
-    width: SCREEN_WIDTH,
-    resizeMode: 'cover'
-  },
-  content: {
-    shadowColor: '#222',
-    shadowOpacity: 0.3,
-    shadowRadius: 2,
-    backgroundColor: '#fff',
-    flex: 1,
-    flexDirection: 'column'
-  },
-  headerView: {
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  avatarView: {
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  listView: {
-    backgroundColor: 'rgba(247,247, 250, 1)'
-  },
-  logoutText: {
-    color: 'red',
-    textAlign: 'center',
-    fontWeight: 'bold'
-  }
-});
